♻️refactor: mealPhotoStore import 경로를 utils/store로 수정
